level 1 problem 3

+++ Problems/Level-1/Problem_3/manasareddy_problems_L1_3.c
@@ -4,26 +4,27 @@ int main()
 {
     int N;
     scanf("%d",&N);
-    double k,j;
-    scanf("%lf",&k);
-    scanf("%lf",&j);
+    float k,j;
+    scanf("%f",&k);
+    scanf("%f",&j);
     int arr[N];
     for(int i=0; i<N; i++)
     {
         scanf("%d",&arr[i]);
     }
-    double sum1=0;
+    float sum1=0;
     for(int i=0; i<k;i++)
     {
         sum1=sum1+arr[i];
     }
-    double mean1= sum1/k;
+    float mean1= sum1/k;
     int sum2=0;
     for(int i=N-j;  i<N;i ++)
     {
         sum2=sum2+arr[i];
     }
-    double mean2= sum2/j;
-    double final= mean1-mean2;
-    printf("%lf",final);
+    float mean2= sum2/j;
+    float final= mean1-mean2;
+    printf("%0.1f",final);
+
 }
